pass tile size as double

+++ lib/maplibre-compose/src/iosMain/kotlin/dev/sargunv/maplibrecompose/core/source/RasterSource.kt
@@ -6,5 +6,5 @@ import platform.Foundation.NSURL
 public actual class RasterSource actual constructor(id: String, configUrl: String, tileSize: Int) :
   Source() {
   override val impl: MLNRasterTileSource =
-    MLNRasterTileSource(id, NSURL(string = configUrl), tileSize)
+    MLNRasterTileSource(id, NSURL(string = configUrl), tileSize.toDouble())
 }
